refactor(client): extract mobile, admin-db, and files APIs from lib/api.ts
Move feature-specific API code into dedicated modules under api/,
reducing lib/api.ts from 369 to 82 lines (core infra only).
