+++ client/src/lib/api.ts
@@ -1,135 +1,3 @@
-// --- Memoized API Request Utility ---
-type CacheEntry = { data: any; expires: number };
-export const apiCache = new Map<string, CacheEntry>();
-const DEFAULT_TTL = 60 * 1000; // 60 Sekunden
-
-export async function memoizedApiRequest<T = any>(url: string, params?: any, ttl: number = DEFAULT_TTL): Promise<T> {
-  const key = url + JSON.stringify(params || {});
-  const now = Date.now();
-  const cached = apiCache.get(key);
-  if (cached && cached.expires > now) {
-    return cached.data;
-  }
-  const res = await apiClient.get(url, { params });
-  apiCache.set(key, { data: res.data, expires: now + ttl });
-  return res.data;
-}
-
-// --- Mobile Devices API ---
-export interface MobileRegistrationToken {
-  token: string;
-  server_url: string;
-  expires_at: string;
-  qr_code: string;
-  vpn_config?: string;
-  device_token_validity_days: number;
-}
-
-export interface MobileDevice {
-  id: string;
-  user_id: number;  // Changed from string to number
-  username?: string; // Nur für Admin sichtbar
-  device_name: string;
-  device_type: string;
-  device_model: string | null;
-  os_version: string | null;
-  app_version: string | null;
-  is_active: boolean;
-  last_sync: string | null;
-  last_seen?: string | null;
-  expires_at: string | null;
-  created_at: string;
-  updated_at: string | null;
-}
-
-export async function generateMobileToken(
-  includeVpn: boolean = false,
-  deviceName: string = 'Mobile Device',
-  tokenValidityDays: number = 90,
-  vpnType: string = 'auto'
-): Promise<MobileRegistrationToken> {
-  const res = await apiClient.post('/api/mobile/token/generate', null, {
-    params: {
-      include_vpn: includeVpn,
-      device_name: deviceName,
-      token_validity_days: tokenValidityDays,
-      vpn_type: vpnType
-    }
-  });
-  return res.data;
-}
-
-export async function getAvailableVpnTypes(): Promise<string[]> {
-  const res = await apiClient.get('/api/vpn/available-types');
-  return res.data.available_types;
-}
-
-export async function getMobileDevices(): Promise<MobileDevice[]> {
-  // Add cache-busting timestamp to prevent stale data
-  const res = await apiClient.get('/api/mobile/devices', {
-    params: { _t: Date.now() }
-  });
-  return res.data;
-}
-
-export async function deleteMobileDevice(deviceId: string): Promise<void> {
-  await apiClient.delete(`/api/mobile/devices/${deviceId}`);
-}
-
-export interface ExpirationNotification {
-  id: string;
-  notification_type: string;
-  sent_at: string;
-  success: boolean;
-  error_message: string | null;
-  device_expires_at: string | null;
-}
-
-export async function getDeviceNotifications(deviceId: string, limit: number = 10): Promise<ExpirationNotification[]> {
-  const res = await apiClient.get(`/api/mobile/devices/${deviceId}/notifications`, {
-    params: { limit }
-  });
-  return res.data;
-}
-
-// --- Duplicate Check API ---
-export interface ExistingFileInfo {
-  filename: string;
-  size_bytes: number;
-  modified_at: string;
-  checksum: string | null;
-}
-
-export async function checkFilesExist(
-  filenames: string[],
-  targetPath: string,
-): Promise<{ duplicates: ExistingFileInfo[] }> {
-  const res = await apiClient.post('/api/files/check-exists', {
-    filenames,
-    target_path: targetPath,
-  });
-  return res.data;
-}
-
-// --- File Permissions API ---
-export async function getFilePermissions(path: string) {
-  // Memoized GET: Permissions werden für 60s gecached
-  return memoizedApiRequest(`/api/files/permissions`, { path });
-}
-
-export async function setFilePermissions(data: {
-  path: string;
-  owner_id: number;
-  rules: Array<{
-    user_id: number;
-    can_view: boolean;
-    can_edit: boolean;
-    can_delete: boolean;
-  }>;
-}) {
-  const res = await apiClient.put(`/api/files/permissions`, data);
-  return res.data;
-}
 import axios from 'axios';
 
 export const API_VERSION = '1';
@@ -195,108 +63,20 @@ apiClient.interceptors.response.use(
   }
 );
 
-// --- Admin DB API (read-only, admin-only) ---
-export interface AdminTableSchemaField {
-  name: string;
-  type: string;
-  nullable: boolean;
-  default?: any;
-}
+// --- Memoized API Request Utility ---
+type CacheEntry = { data: any; expires: number };
+export const apiCache = new Map<string, CacheEntry>();
+const DEFAULT_TTL = 60 * 1000; // 60 Sekunden
 
-export interface AdminTableSchemaResponse {
-  table: string;
-  columns: AdminTableSchemaField[];
-}
-
-export interface AdminTableRowsResponse {
-  table: string;
-  page: number;
-  page_size: number;
-  rows: Array<Record<string, any>>;
-  total?: number | null;
-  sort_by?: string | null;
-  sort_order?: string | null;
-}
-
-export interface AdminTableCategoriesResponse {
-  categories: Record<string, string[]>;
-}
-
-export interface ColumnFilter {
-  op: 'contains' | 'eq' | 'gt' | 'lt' | 'gte' | 'lte' | 'between' | 'is_null' | 'is_true' | 'is_false';
-  value?: string | number;
-  from?: string | number;
-  to?: string | number;
-}
-
-export type ColumnFilters = Record<string, ColumnFilter>;
-
-export async function getAdminTables(): Promise<string[]> {
-  const res = await apiClient.get('/api/admin/db/tables');
-  return res.data.tables;
-}
-
-export async function getAdminTableCategories(): Promise<AdminTableCategoriesResponse> {
-  const res = await apiClient.get('/api/admin/db/tables/categories');
-  return res.data;
-}
-
-export async function getAdminTableSchema(table: string): Promise<AdminTableSchemaResponse> {
-  const res = await apiClient.get(`/api/admin/db/table/${encodeURIComponent(table)}/schema`);
-  return res.data;
-}
-
-export async function getAdminTableRows(
-  table: string,
-  page: number = 1,
-  page_size: number = 50,
-  fields?: string[],
-  q?: string,
-  sort_by?: string,
-  sort_order?: string,
-  filters?: ColumnFilters
-): Promise<AdminTableRowsResponse> {
-  const params: any = { page, page_size };
-  if (fields && fields.length) params.fields = fields.join(',');
-  if (q) params.q = q;
-  if (sort_by) params.sort_by = sort_by;
-  if (sort_order) params.sort_order = sort_order;
-  if (filters && Object.keys(filters).length > 0) params.filters = JSON.stringify(filters);
-  const res = await apiClient.get(`/api/admin/db/table/${encodeURIComponent(table)}`, { params });
-  return res.data;
-}
-
-// --- Database Health & Info ---
-export interface DatabaseHealthResponse {
-  is_healthy: boolean;
-  connection_status: string;
-  database_type: string;
-  integrity_check?: string;
-  pool_size?: number;
-  pool_checked_in?: number;
-  pool_checked_out?: number;
-  pool_overflow?: number;
-}
-
-export interface TableSizeInfo {
-  table_name: string;
-  row_count: number;
-  estimated_size_bytes: number;
-}
-
-export interface DatabaseInfoResponse {
-  database_type: string;
-  total_size_bytes: number;
-  tables: TableSizeInfo[];
-}
-
-export async function getDatabaseHealth(): Promise<DatabaseHealthResponse> {
-  const res = await apiClient.get('/api/admin/db/health');
-  return res.data;
-}
-
-export async function getDatabaseInfo(): Promise<DatabaseInfoResponse> {
-  const res = await apiClient.get('/api/admin/db/info');
+export async function memoizedApiRequest<T = any>(url: string, params?: any, ttl: number = DEFAULT_TTL): Promise<T> {
+  const key = url + JSON.stringify(params || {});
+  const now = Date.now();
+  const cached = apiCache.get(key);
+  if (cached && cached.expires > now) {
+    return cached.data;
+  }
+  const res = await apiClient.get(url, { params });
+  apiCache.set(key, { data: res.data, expires: now + ttl });
   return res.data;
 }
 
@@ -306,64 +86,4 @@ export function extractErrorMessage(detail: unknown, fallback: string): string {
     return detail.map((e: any) => e.msg ?? String(e)).join('; ');
   }
   return fallback;
-}
-
-
-// ============================================================================
-// Ownership Transfer API
-// ============================================================================
-
-export interface ConflictInfo {
-  original_path: string;
-  resolved_path: string | null;
-  action: string;
-}
-
-export interface OwnershipTransferRequest {
-  path: string;
-  new_owner_id: number;
-  recursive?: boolean;
-  conflict_strategy?: 'rename' | 'skip' | 'overwrite';
-}
-
-export interface OwnershipTransferResponse {
-  success: boolean;
-  message: string;
-  transferred_count: number;
-  skipped_count: number;
-  new_path: string | null;
-  conflicts: ConflictInfo[];
-  error: string | null;
-}
-
-export async function transferOwnership(
-  request: OwnershipTransferRequest
-): Promise<OwnershipTransferResponse> {
-  const res = await apiClient.post('/api/files/transfer-ownership', request);
-  return res.data;
-}
-
-export interface ResidencyViolation {
-  path: string;
-  current_owner_id: number;
-  current_owner_username: string;
-  expected_directory: string;
-  actual_directory: string;
-}
-
-export interface EnforceResidencyRequest {
-  dry_run?: boolean;
-  scope?: string | null;
-}
-
-export interface EnforceResidencyResponse {
-  violations: ResidencyViolation[];
-  fixed_count: number;
-}
-
-export async function enforceResidency(
-  request: EnforceResidencyRequest
-): Promise<EnforceResidencyResponse> {
-  const res = await apiClient.post('/api/files/enforce-residency', request);
-  return res.data;
 }
