+++ client/src/pages/AdminDatabase.tsx
@@ -8,7 +8,7 @@ import TableSidebar from '../components/admin/TableSidebar'
 import RowDetailPanel from '../components/admin/RowDetailPanel'
 import DataTypeIndicator from '../components/admin/DataTypeIndicator'
 import { rowsToCsv } from '../lib/csv'
-import type { ColumnFilters, AdminTableSchemaField } from '../lib/api'
+import type { ColumnFilters, AdminTableSchemaField } from '../api/admin-db'
 import { Database, Table, Download, ChevronLeft, ChevronRight, RefreshCw, BarChart3, History, Wrench, Search, Filter, X } from 'lucide-react'
 import { AdminBadge } from '../components/ui/AdminBadge'
 import DatabaseStatsCards from '../components/admin/DatabaseStatsCards'
